Create report_clock
This code is a graphical alarm clock. In the 10 minutes left to the synop or metar or half hour metar, the clock rings. It depends on your settings. It creates a text file with itself to write the settings into it to use when the program is restarted. Adjustments are made possible by the radio buttons. This clock also works in a system tray form.